+++ settings.gradle
@@ -1,2 +1,2 @@
+rootProject.name = "SnipSnap"
 include ':app'
-rootProject.name = "Snipsnap"
